Reorganize and modernize CMakeLists.txt
Also fix coverage with Clang and STM32 build
