+++ CMakeLists.txt
@@ -18,7 +18,7 @@
 # SPDX-License-Identifier: Apache-2.0 OR LGPL-2.1-or-later
 #
 
-cmake_minimum_required (VERSION 2.6)
+cmake_minimum_required (VERSION 3.13)
 
 project(AtomVM)
 set(CMAKE_MODULE_PATH ${PROJECT_SOURCE_DIR}/CMakeModules)
@@ -26,30 +26,22 @@ set(CMAKE_MODULE_PATH ${PROJECT_SOURCE_DIR}/CMakeModules)
 find_package(Elixir)
 
 option(AVM_DISABLE_FP "Disable floating point support." OFF)
-if (AVM_DISABLE_FP)
-    add_definitions(-DAVM_NO_FP)
-endif()
-
 option(AVM_USE_32BIT_FLOAT "Use 32 bit floats." OFF)
-if (AVM_USE_32BIT_FLOAT)
-    add_definitions(-DAVM_USE_SINGLE_PRECISION)
-endif()
-
 option(AVM_VERBOSE_ABORT "Print module and line number on VM abort" OFF)
-if (AVM_VERBOSE_ABORT)
-    add_definitions(-DAVM_VERBOSE_ABORT)
+option(COVERAGE "Build for code coverage" OFF)
+
+if((${CMAKE_SYSTEM_NAME} STREQUAL "Darwin") OR
+   (${CMAKE_SYSTEM_NAME} STREQUAL "Linux") OR
+   (${CMAKE_SYSTEM_NAME} STREQUAL "FreeBSD"))
+    add_subdirectory(src/platforms/generic_unix)
+else()
+    message(FATAL_ERROR "Unsupported platform: ${CMAKE_SYSTEM_NAME}")
 endif()
 
-add_subdirectory(src)
 add_subdirectory(tests)
 add_subdirectory(tools/packbeam)
 if (NOT "${CMAKE_GENERATOR}" MATCHES "Xcode")
     add_subdirectory(libs)
     add_subdirectory(examples)
     add_subdirectory(doc)
-endif()
-
-if (CMAKE_BUILD_TYPE STREQUAL "Coverage")
-    include(CodeCoverage)
-    SETUP_TARGET_FOR_COVERAGE(NAME coverage EXECUTABLE test-erlang DEPENDENCIES test-erlang)
 endif()
