--- src/CMakeLists.txt
@@ -1,71 +1,0 @@
-#
-# This file is part of AtomVM.
-#
-# Copyright 2017-2020 Davide Bettio <davide@uninstall.it>
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# SPDX-License-Identifier: Apache-2.0 OR LGPL-2.1-or-later
-#
-
-cmake_minimum_required (VERSION 2.6)
-project (AtomVM)
-
-option(ADVANCED_TRACING "Enable advanced tracing" OFF)
-if (ADVANCED_TRACING)
-    add_definitions(-DENABLE_ADVANCED_TRACE)
-endif()
-
-add_subdirectory(libAtomVM)
-
-if((${CMAKE_SYSTEM_NAME} STREQUAL "Darwin") OR
-   (${CMAKE_SYSTEM_NAME} STREQUAL "Linux") OR
-   (${CMAKE_SYSTEM_NAME} STREQUAL "FreeBSD"))
-    add_subdirectory(platforms/generic_unix)
-else()
-    message(FATAL_ERROR "Unsupported platform: ${CMAKE_SYSTEM_NAME}")
-endif()
-
-if(${CMAKE_SYSTEM_NAME} STREQUAL "Linux")
-    include(CheckFunctionExists)
-    include(CheckLibraryExists)
-    check_library_exists(rt clock_gettime "time.h" HAVE_CLOCK_GETTIME)
-    if (HAVE_CLOCK_GETTIME)
-        set(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} -lrt")
-        set(CMAKE_C_STANDARD_LIBRARIES "${CMAKE_C_STANDARD_LIBRARIES} -lrt")
-    else()
-        # might also be in libc
-        check_library_exists(c clock_gettime "" HAVE_CLOCK_GETTIME)
-    endif()
-endif()
-
-set(
-    PLATFORM_LIB_SUFFIX
-    ${CMAKE_SYSTEM_NAME}-${CMAKE_SYSTEM_PROCESSOR}
-)
-
-if(CMAKE_COMPILER_IS_GNUCC)
-    set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wall -pedantic -Wextra -ggdb")
-endif()
-
-include_directories(${CMAKE_CURRENT_BINARY_DIR} libAtomVM/)
-
-add_executable(AtomVM main.c)
-target_link_libraries(AtomVM libAtomVM libAtomVM${PLATFORM_LIB_SUFFIX})
-set_property(TARGET AtomVM PROPERTY C_STANDARD 99)
-
-if (CMAKE_BUILD_TYPE STREQUAL "Coverage")
-    set_target_properties(AtomVM PROPERTIES COMPILE_FLAGS "-O0 -fprofile-arcs -ftest-coverage")
-endif()
-
-install(TARGETS AtomVM DESTINATION lib/atomvm)
