+++ src/platforms/esp32/components/avm_builtins/CMakeLists.txt
@@ -38,3 +38,5 @@ idf_component_register(
 idf_build_set_property(
     LINK_OPTIONS "-Wl,--whole-archive ${CMAKE_CURRENT_BINARY_DIR}/lib${COMPONENT_NAME}.a -Wl,--no-whole-archive"
     APPEND)
+
+target_compile_features(${COMPONENT_LIB} INTERFACE c_std_11)
