+++ src/platforms/generic_unix/lib/platform_defaultatoms.c
