+++ src/platforms/generic_unix/lib/sys.c
